+++ apicast/src/proxy.lua
@@ -9,6 +9,7 @@
 local env = require 'resty.env'
 local custom_config = env.get('APICAST_CUSTOM_CONFIG')
 local configuration_store = require 'configuration_store'
+local util = require('util')
 
 local oauth = require 'oauth'
 local resty_url = require 'resty.url'
@@ -22,12 +23,21 @@ local concat = table.concat
 local gsub = string.gsub
 local tonumber = tonumber
 local setmetatable = setmetatable
-local exit = ngx.exit
 local encode_args = ngx.encode_args
 local resty_resolver = require 'resty.resolver'
+local semaphore = require('ngx.semaphore')
+local backend_client = require('backend_client')
+local timers = semaphore.new(tonumber(env.get('APICAST_REPORTING_THREADS') or 0))
+
 local empty = {}
 
 local response_codes = env.enabled('APICAST_RESPONSE_CODES')
+
+local using_post_action = response_codes or timers:count() < 1
+
+if not using_post_action then
+  ngx.log(ngx.WARN, 'using experimental asynchronous reporting threads: ', timers:count())
+end
 
 local _M = { }
 
@@ -143,7 +153,7 @@ local http = {
     local backend_upstream = ngx.ctx.backend_upstream
     local previous_real_url = ngx.var.real_url
     ngx.log(ngx.DEBUG, '[ctx] copying backend_upstream of size: ', #backend_upstream)
-    local res = ngx.location.capture(assert(url), { share_all_vars = true, ctx = { backend_upstream = backend_upstream } })
+    local res = ngx.location.capture(assert(url), { share_all_vars = true, ctx = { backend_upstream = backend_upstream, backend_endpoint = ngx.var.backend_endpoint } })
 
     local real_url = ngx.var.real_url
 
@@ -191,6 +201,8 @@ function _M:authorize(service, usage, credentials)
     ngx.log(ngx.DEBUG, 'apicast cache hit key: ', cached_key)
     ngx.var.cached_key = cached_key
   else
+
+    self:set_backend_upstream(service)
     ngx.log(ngx.INFO, 'apicast cache miss key: ', cached_key)
     local res = http.get(internal_location)
 
@@ -207,6 +219,10 @@ function _M:authorize(service, usage, credentials)
     end
     -- set cached_key to nil to avoid doing the authrep in post_action
     ngx.var.cached_key = nil
+  end
+
+  if not using_post_action then
+    self:post_action(true)
   end
 end
 
@@ -281,8 +297,6 @@ function _M:call(host)
   host = host or ngx.var.host
   local service = ngx.ctx.service or self:set_service(host)
 
-  self:set_backend_upstream(service)
-
   if service.backend_version == 'oauth' then
     local o = oauth.new(self.configuration)
     local f, params = oauth.call(o, service)
@@ -347,22 +361,77 @@ function _M:access(service)
   return self:authorize(service, usage, credentials)
 end
 
-
-local function response_codes_encoded_data()
+local function response_codes_data()
   local params = {}
 
   if not response_codes then
-    return ''
+    return params
   end
 
   if response_codes then
     params["log[code]"] = ngx.var.status
   end
 
-  return ngx.escape_uri(ngx.encode_args(params))
+  return params
 end
 
-function _M:post_action()
+local function response_codes_encoded_data()
+  return ngx.escape_uri(ngx.encode_args(response_codes_data()))
+end
+
+local function handle_post_action_response(cached_key, res)
+  if res.ok == false or res.status ~= 200 then
+    local api_keys = ngx.shared.api_keys
+
+    if api_keys then
+      ngx.log(ngx.NOTICE, 'apicast cache delete key: ', cached_key, ' cause status ', res.status)
+      api_keys:delete(cached_key)
+    else
+      ngx.log(ngx.ALERT, 'apicast cache error missing shared memory zone api_keys')
+    end
+
+    ngx.log(ngx.ERR, 'http_client error: ', res.error, ' status: ', res.status)
+  end
+end
+
+local function post_action(_, cached_key, backend, ...)
+  local res = util.timer('backend post_action', backend.authrep, backend, ...)
+
+  if not using_post_action then
+    timers:post(1)
+  end
+
+  handle_post_action_response(cached_key, res)
+end
+
+local function capture_post_action(self, cached_key, service)
+  self:set_backend_upstream(service)
+
+  local auth_uri = service.backend_version == 'oauth' and 'threescale_oauth_authrep' or 'threescale_authrep'
+  local res = http.get("/".. auth_uri .."?log=" .. response_codes_encoded_data())
+
+  handle_post_action_response(cached_key, res)
+end
+
+local function timer_post_action(self, cached_key, service)
+  local backend = assert(backend_client:new(service), 'missing backend')
+
+  local ok, err = timers:wait(10)
+
+  if ok then
+    -- TODO: try to do this in different phase and use semaphore to limit number of background threads
+    -- TODO: Also it is possible to use sets in shared memory to enqueue work
+    ngx.timer.at(0, post_action, cached_key, backend, ngx.ctx.usage, ngx.ctx.credentials, response_codes_data())
+  else
+    ngx.log(ngx.ERR, 'failed to acquire timer: ', err)
+    return capture_post_action(self, cached_key, service)
+  end
+end
+
+function _M:post_action(force)
+  if not using_post_action and not force then
+    return nil, 'post action not needed'
+  end
 
   local cached_key = ngx.var.cached_key
 
@@ -371,26 +440,15 @@ function _M:post_action()
 
     local service_id = ngx.var.service_id
     local service = ngx.ctx.service or self.configuration:find_by_id(service_id)
-    self:set_backend_upstream(service)
 
-    local auth_uri = service.backend_version == 'oauth' and 'threescale_oauth_authrep' or 'threescale_authrep'
-    local res = http.get("/".. auth_uri .."?log=" .. response_codes_encoded_data())
-
-    if res.status ~= 200 then
-      local api_keys = ngx.shared.api_keys
-
-      if api_keys then
-        ngx.log(ngx.NOTICE, 'apicast cache delete key: ', cached_key, ' cause status ', res.status)
-        api_keys:delete(cached_key)
-      else
-        ngx.log(ngx.ALERT, 'apicast cache error missing shared memory zone api_keys')
-      end
+    if using_post_action then
+      capture_post_action(self, cached_key, service)
+    else
+      timer_post_action(self, cached_key, service)
     end
   else
     ngx.log(ngx.INFO, '[async] skipping after action, no cached key')
   end
-
-  exit(ngx.HTTP_OK)
 end
 
 if custom_config then