+++ t/003-apicast.t
@@ -469,6 +469,7 @@ env RESOLVER=127.0.0.1:1953;
 --- http_config
   include $TEST_NGINX_UPSTREAM_CONFIG;
   lua_package_path "$TEST_NGINX_LUA_PATH";
+  lua_shared_dict api_keys 1m;
   init_by_lua_block {
     require('configuration_loader').mock({
       services = {
@@ -502,6 +503,7 @@ env RESOLVER=127.0.0.1:1953;
        if ngx.var.host == 'localhost.example.com' then
          ngx.exit(200)
        else
+         ngx.log(ngx.ERR, 'invalid host: ', ngx.var.host)
          ngx.exit(404)
        end
      }